Don't require ALPN
Since version 1.67 of the Go gRPC package ALPN is mandatory. But the
OpenShift router doesn't seem to support it. To address that we need to
use an experimental version of the credentials package that doesn't
require ALPN.

This should be addressed in the future, ideally configuring OpenShift so
that it does support ALPN, as the experiemental credentials package will
be eventually removed.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>