Add support for login with token generation command
This patch adds a new `--token-command` flag to the `login` commands
that uses a command provided by the user to generate the authentication
tokens. For example, this would automatically request/renew the token of
a Kubernetes service account:

```shell
$ fulfillment-clin login \
--token-command 'kubectl create token -n innabox client --duration 1h` \
--address ...
```

If the generated token is a JWT the tool will check the expiry time and
run the command only when the token has expired. If it isn't a JWT the
command will be executed every time that a token is needed.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>