Add support for login with token generation script (#8)
This patch adds a new `--token-script` flag to the `login` commands that
uses a shell command provided by the user to generate the authentication
tokens. For example, this would automatically request/renew the token of
a Kubernetes service account:

```shell
$ fulfillment-cli login \
--token-script 'kubectl create token -n innabox client --duration 1h` \
--address ...
```

The shell command will be passed to the shell of the user for execution,
something similar to `bash -c '...'`, and the result written to the
standard output will be the token.

If the shell command returns a non zero exit code it will be considered
an error and the token will be discarded.

If the generated token is a JWT the tool will check the expiry time and
run the shell command only when the token has expired. If it isn't a JWT
the command will be executed every time that a token is needed.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>