Generic `get` command (#11)
* Generic `get` command

Currently we have different implementations of the `get` command for
different types of objects. This patch replaces those with a single
implementation that supports all existing and future object types.

This generic `get` command is implemented using protocol buffers
reflection, taking advantage of the fact that all our objects and
services use the same fields and methods. For example, all the services
have a `List` method to get multiple objects, and the response object
for that method always has an `items` field that contains the results.

The rendering of the results is drive by a table description that is in
a YAML file. For example, the table for the `clusterorder` object is
this:

```yaml
full_name: fulfillment.v1.ClusterOrder
columns:

- header: ID
  value: id

- header: TEMPLATE
  value: spec.template_id

- header: STATE
  value: status.state
  type: fulfillment.v1.ClusterOrderState

- header: CLUSTER
  value: status.cluster_id
```

The `full_name` is the full name of the protocol buffers type, including
the package prefix.

For each column the `header` is the value displayed in the header of the
table.

The `value` is the CEL expression used to calculate the text to display.

The `type` is used to indicate the type of the result of the CEL
expression, and is used in particular for enumerated types to shorten
the displayed value. For example, if the value is
`CLUSTER_ORDER_STATE_FULFILLED` the displayed value will be just
`FULFILLED`.

For example, the output of the `get clusterorders` command will look
like this:

```
ID                                    TEMPLATE        STATE        CLUSTER
6209c13d-4836-4908-a2b0-e6294e3ad833  ocp_4_17_small  FULFILLED    97463a75-439f-4610-9d13-542f1ad72b7e
```

The new generic command also supports JSON and YAML format, with the
`--output json` or `--output yaml` flags. For example the `get
clustertemplates -o json` command will generate the following JSON
output:

```json
{
  "@type": "type.googleapis.com/fulfillment.v1.ClusterTemplate",
  "description": "OpenShift 4.17 with `small` instances as worker nodes.",
  "id": "ocp_4_17_small",
  "metadata": {
    "creation_timestamp": "2025-04-16T16:54:25.827610Z"
  },
  "parameters": [
    {
      "default": {
        "@type": "type.googleapis.com/google.protobuf.BoolValue",
        "value": true
      },
      "name": "my_bool",
      "type": "type.googleapis.com/google.protobuf.BoolValue"
    },
    {
      "default": {
        "@type": "type.googleapis.com/google.protobuf.Int32Value",
        "value": 42
      },
      "name": "my_int",
      "type": "type.googleapis.com/google.protobuf.Int32Value"
    },
    {
      "default": {
        "@type": "type.googleapis.com/google.protobuf.StringValue",
        "value": "my_value"
      },
      "name": "my_string",
      "type": "type.googleapis.com/google.protobuf.StringValue"
    }
  ],
  "title": "OpenShift 4.17 small"
}
```

And the `get clustertemplates -o yaml` will generate the following:

```yaml
'@type': type.googleapis.com/fulfillment.v1.ClusterTemplate
description: OpenShift 4.17 with `small` instances as worker nodes.
id: ocp_4_17_small
metadata:
  creation_timestamp: "2025-04-16T16:54:25.827610Z"
parameters:
  - default:
      '@type': type.googleapis.com/google.protobuf.BoolValue
      value: true
    name: my_bool
    type: type.googleapis.com/google.protobuf.BoolValue
  - default:
      '@type': type.googleapis.com/google.protobuf.Int32Value
      value: 42
    name: my_int
    type: type.googleapis.com/google.protobuf.Int32Value
  - default:
      '@type': type.googleapis.com/google.protobuf.StringValue
      value: my_value
    name: my_string
    type: type.googleapis.com/google.protobuf.StringValue
title: OpenShift 4.17 small
```

Note that both the JSON and YAML outputs automaticall add the `@type`
field to the representation of the object.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

* Add API and console URL to clusters table

This patch restores the API and console URL fields to the clusters table
that were removed by accident by a previous commit. For example:

```shell
$ fulfillment-cli get clusters
ID                                    STATE  API URL         CONSOLE URL
7501325a-3b15-4cdb-ad1e-3822babc3eab  READY  https://my.api  https://my.console
```

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

* Show list of available object types

Currently when the `get` command is executed without arguments it prints
the following error message:

```
Error: requires at least 1 arg(s), only received 0
```

It used to print more useful information before the introduction of the
generic implementation.

This patch improves the output, so that it will explain that the object
type is mandatory and will also list the available object types:

```shell
$ fulfillment-cli get
You must specify the type of object to get.

The following object types are available:

- clusterorders
- clusters
- clustertemplates

For example, to get the list of cluster templates:

./fulfillment-cli get clustertemplates

Use the '--help' option to get more details about the command.
```

That list of object types is obtained scanning the protobuf type
registry, so it will be automatically updated when we add new types in
the future.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

---------

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>