Add edit command
This patch adds a new `edit` command that can be used to edit objects of
any of the supported types. The command gets the current representation
of the object, writes it to a temporary file, invokes an editor to edit
it and then, when the user writes the result and closes the editor, it
tries to save the result.

A good part of the code required for this new command is also required
for the `get` command. To avoid duplicating that this patch also moves
that to a reusable _ReflectionHelper_ type.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>