Don't show deleted objects by default (#17)
* Update to version 0.0.7 of the API

The more relevant changes in the new versionof the API are the
following:

- Add the `node_sets` field to clusters.
- Add the `HostClass` type and service.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

* Add edit command

This patch adds a new `edit` command that can be used to edit objects of
any of the supported types. The command gets the current representation
of the object, writes it to a temporary file, invokes an editor to edit
it and then, when the user writes the result and closes the editor, it
tries to save the result.

A good part of the code required for this new command is also required
for the `get` command. To avoid duplicating that this patch also moves
that to a reusable _ReflectionHelper_ type.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

* Add `--filter` flag to the `get` command

This patch adds a new `--filter` flag to the `get` command. The filter
is a CEL expression that the server will use to decide which results to
send.

The object will be returned only if the result of the expression is
`true`.

The `this` variable in this CEL expression refers to the object. Fields
of the object can be used with this as a prefix.

For example, to find all the cluster templates that have a title
starting with `Open`:

```shell
$ ./fulfillment-cli get clustertemplates --filter "this.title.startsWith('Open')"
ID              TITLE                 DESCRIPTION
ocp_4_17_small  OpenShift 4.17 small  OpenShift 4.17 with `small` instances as worker nodes.
```

Many of the CEL operators and built-in functions are available, but not
100% of them. Documentation about that will be added to the server in
the future.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

* Don't show deleted objects by default

This patch changes the `get` command so that by default it doesn't show
deleted objects. It also adds a new `--include-deleted` flag that
changes that behaviour.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

---------

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>